# internal/adapters/postgresql/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package repo

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type CartItem struct {
	ID        int32 `json:"id"`
	UserID    int64 `json:"user_id"`
	ProductID int64 `json:"product_id"`
	Quantity  int32 `json:"quantity"`
}

type Order struct {
	ID        int32              `json:"id"`
	UserID    int64              `json:"user_id"`
	Total     pgtype.Numeric     `json:"total"`
	Status    string             `json:"status"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type OrderItem struct {
	ID        int32          `json:"id"`
	OrderID   int32          `json:"order_id"`
	ProductID int64          `json:"product_id"`
	Quantity  int32          `json:"quantity"`
	Price     pgtype.Numeric `json:"price"`
}

type Product struct {
	ID           int64            `json:"id"`
	Name         string           `json:"name"`
	PriceInCents int32            `json:"price_in_cents"`
	Quantity     int32            `json:"quantity"`
	CreatedAt    pgtype.Timestamp `json:"created_at"`
}

type User struct {
	ID           int64            `json:"id"`
	Email        string           `json:"email"`
	PasswordHash string           `json:"password_hash"`
	Name         string           `json:"name"`
	Role         string           `json:"role"`
	CreatedAt    pgtype.Timestamp `json:"created_at"`
}